Updated script.js The updateDisplay function was removed and substituted with update of the display using the values stored in the variables. Function to operate introduced. Introduced logic to handle calculations and for the buttons to clear and delete the last input.
The calculator at this stage is functional, despite having the % button not working yet.
Improvements can be made, on the readability of the code, e.g., using separate event handlers for different group of buttons.
The choice of event delegation to the buttons container might not be the best in this example.